refactor(search-box): move interfaces
